Support Python 3.11 (#5994)
* remove some more deprecated numpy stuff

* workaround for changes to random.shuffle in python 3.11

* fix pickling of rdkit mols in python 3.11

* add py311 build to CI

* update py311 CI

* remove qt for py311 for the moment

* only use the new code with pyversion >=3.11

* use the new logic for all pickle_suites

* need to work with older py too

+++ Code/GraphMol/Depictor/Wrap/testDepictor.py
@@ -28,7 +28,7 @@ def getDistMat(mol):
     conf = mol.GetConformer()
     nat = mol.GetNumAtoms()
     nl = nat * (nat - 1) // 2
-    res = np.zeros(nl, np.float)
+    res = np.zeros(nl, float)
 
     for i in range(1, nat):
         pi = conf.GetAtomPosition(i)
